requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
black==24.3.0
cellpose @ git+https://www.github.com/mouseland/cellpose.git@52f75f9636250979dcff48da18d24aaff41c2c86
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
COCO_merger==0.0.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
-e git+https://github.com/facebookresearch/detectron2.git@2a420edb307c9bdf640f036d3b196bed474b8593#egg=detectron2
fastremap==1.15.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
fvcore==0.1.5.post20221221
grpcio==1.62.1
h5py==3.11.0
hydra-core==1.3.2
idna==3.7
imagecodecs==2024.9.22
imageio==2.34.1
imageio-ffmpeg==0.5.1
imglyb==2.1.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.9
jgo==1.0.6
Jinja2==3.1.3
joblib==1.4.2
JPype1==1.5.0
kiwisolver==1.4.5
labeling==0.1.14
lazy_loader==0.4
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
proglog==0.1.10
protobuf==5.26.1
psutil==6.0.0
pycocotools==2.0.7
pyimagej==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.1
requests==2.32.3
roifile==2024.9.15
scikit-image==0.22.0
scipy==1.13.0
scyjava==1.10.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
texttable==1.7.0
tifffile==2024.4.18
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.2
xarray==2024.7.0
yacs==0.1.8
zipp==3.18.1
